refactor: #82 커게시물 목록 페이지 컴포넌트 분리
